requirements.txt
customtkinter
python-dotenv
Pillow
asyncio
imap_tools
icecream
imaplib
Email
PyInstaller